requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
tensorflow==2.15.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
pillow==10.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
httpx==0.26.0
